+++ DX11Renderer/src/Renderer/InputLayout.cpp
@@ -2,10 +2,10 @@
 
 namespace Yassin
 {
-	InputLayout::InputLayout(ID3DBlob* pBlob, const D3D11_INPUT_ELEMENT_DESC* ied, unsigned int size)
+	InputLayout::InputLayout(ID3DBlob* pBlob, const D3D11_INPUT_ELEMENT_DESC* ied, size_t size)
 	{
 		HRESULT hr;	
-		hr = RendererContext::GetDevice()->CreateInputLayout(ied, size, pBlob->GetBufferPointer(), pBlob->GetBufferSize(), &m_InputLayout);
+		hr = RendererContext::GetDevice()->CreateInputLayout(ied, (unsigned int)size, pBlob->GetBufferPointer(), pBlob->GetBufferSize(), &m_InputLayout);
 		if(FAILED(hr))
 		{
 			MessageBox(RendererContext::GetWindowHandle(), L"Failed to create input layout", L"Error", MB_OK);
