+++ DX11Renderer/src/Renderer/InputLayout.h
@@ -7,7 +7,7 @@ namespace Yassin
 	class InputLayout
 	{
 	public:
-		InputLayout(ID3DBlob* pBlob, const D3D11_INPUT_ELEMENT_DESC* ied, unsigned int size);
+		InputLayout(ID3DBlob* pBlob, const D3D11_INPUT_ELEMENT_DESC* ied, size_t size);
 		
 		inline void Bind()
 		{
@@ -15,11 +15,11 @@ namespace Yassin
 		}
 
 		const D3D11_INPUT_ELEMENT_DESC* GetIED() { return m_IED; }
-		const unsigned int& GetSize() const { return m_Size; }
+		const size_t& GetSize() const { return m_Size; }
 
 	private:
 		Microsoft::WRL::ComPtr<ID3D11InputLayout> m_InputLayout;
-		unsigned int m_Size;
+		size_t m_Size;
 		const D3D11_INPUT_ELEMENT_DESC* m_IED;
 	};
 }
